+++ gyds-manager/src/main/java/com/fly/index/service/impl/IndexServiceImpl.java
@@ -33,7 +33,7 @@ public class IndexServiceImpl implements IndexService{
 	private CenterDao centerDao;
 
 	@Override
-	public List<AdvertisementDO> getIndexAdvList(Integer areaId,Integer positionNum,Map<String,Object> params) {
+	public List<AdvertisementDO> getIndexAdvList(Long areaId,Integer positionNum,Map<String,Object> params) {
 		List<AdvertisementDO> dataList=new ArrayList<AdvertisementDO>();
 		List<AdvertisementDO> list=new ArrayList<AdvertisementDO>();
 		params.put("positionNum", positionNum);
@@ -222,7 +222,7 @@ public class IndexServiceImpl implements IndexService{
 	}
 
 	@Override
-	public List<AdvertisementDO> getCenterAdvList(Integer regionCode, Integer positionNum) {
+	public List<AdvertisementDO> getCenterAdvList(Long regionCode, Integer positionNum) {
 		List<AdvertisementDO> dataList=new ArrayList<AdvertisementDO>();
 		Map<String,Object> params=new HashMap<>(16);
 		params.put("positionNum", positionNum);
